Loading or selecting unit to product working, code optimizations and modifications save selected unit un posBloc, also modified methods to receibe more params when adding products to POSsale, statefull alert to choose product unit done and working, loading default unit if there is no unit to load, loading correct product price based on unit price, product quantity initialized to unit operation_value, some functions moved from posBloc to providers, and more

+++ lib/models/sale_model.dart
@@ -162,7 +162,7 @@ class SaleModel {
   List<double> netPrice;
   List<double> unitPrice;
   List<double?> realUnitPrice;
-  List<int> quantity;
+  List<double> quantity;
   List<int> productUnit;
   List<int> productUnitIdSelected;
   List<int> productBaseQuantity;
@@ -290,7 +290,7 @@ class SaleModel {
         unitPrice:
             List<double>.from(json["unit_price"].map((x) => x.toDouble())),
         realUnitPrice: List<double>.from(json["real_unit_price"].map((x) => x)),
-        quantity: List<int>.from(json["quantity"].map((x) => x)),
+        quantity: List<double>.from(json["quantity"].map((x) => x)),
         productUnit: List<int>.from(json["product_unit"].map((x) => x)),
         productUnitIdSelected:
             List<int>.from(json["product_unit_id_selected"].map((x) => x)),
